Tests layout: Move tests into game, i18n packages
