+++ src/test/java/soctest/i18n/TestI18NGameoptScenStrings.java
@@ -17,7 +17,7 @@
  *
  * The maintainer of this program can be reached at jsettlers@nand.net
  **/
-package soctest;
+package soctest.i18n;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -320,7 +320,7 @@ public class TestI18NGameoptScenStrings
 
     public static void main(String[] args)
     {
-        org.junit.runner.JUnitCore.main("soctest.TestI18NGameoptScenStrings");
+        org.junit.runner.JUnitCore.main("soctest.i18n.TestI18NGameoptScenStrings");
     }
 
 }
